requirements.txt
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
cssselect==1.0.3
et-xmlfile==1.0.1
idna==2.8
jdcal==1.4.1
lxml==4.3.3
numpy==1.16.4
openpyxl==2.6.2
pandas==0.24.2
Pillow==6.0.0
pyquery==1.4.0
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
selenium==3.141.0
six==1.12.0
soupsieve==1.9.1
urllib3==1.25.3
xpinyin==0.5.6
